Add RegisterTagNameFunc
Added a new function to both the Decoder and Encoder called 'RegisterTagNameFunc'.

NOTES:

Once a custom function has been registered the default, or custom set, tag name is ignored and relies 100% on the function for the name data.

The return value WILL BE CACHED and so return value must be consistent.